finished intermediate Excel
Some of the material was a waste and poorly written. For instance they have a whole section on data parsing at the end, and the way they taught it is impossible to fill everything out. Plus, they don't include all the data to be parsed in the string of code.

Anyway, onto VBA and SAS.

BTW, TIA has effectively made their SAS module only viewable, as they didn't send me any code to download it, and it can't be downloaded outside of an academic course. I will email them about it, but I have had REALLY bad customer service from them the whole way through and DO NOT recommend anyone else using this.